services/backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.110.1
frozenlist==1.3.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
multidict==6.0.4
numpy==1.24.4
openai==0.27.8
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyMySQL==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==4.6.0
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tqdm==4.65.0
typing_extensions==4.11.0
urllib3==2.0.3
uvicorn==0.29.0
yarl==1.9.2
